Move cast_precision_loss to its own module
